Update to hugo 136
Clean up code
introduce self hosted umami
use assets folder for json files
