+++ assets/top-posts.json
@@ -1,0 +1,32 @@
+[
+	{
+		"title": "From requirements.txt to poetry's pyproject.toml",
+		"slug": "requirements-txt-to-poetry-pyproject-toml",
+		"slug_url": "/blog/requirements-txt-to-poetry-pyproject-toml",
+		"path": "blog/2020/20201127-requirements-poetry.md"
+	},
+	{
+		"title": "HTTPX is the new Requests with Async",
+		"slug": "httpx-is-the-new-requests-with-async",
+		"slug_url": "/blog/httpx-is-the-new-requests-with-async",
+		"path": "blog/2020/05072020-httpx-async-client.md"
+	},
+	{
+		"title": "Alpine.js and Tailwind HTML setup",
+		"slug": "Alpine.js and Tailwind HTML setup",
+		"slug_url": "/blog/alpine.js-and-tailwind-html-setup",
+		"path": "blog/2022/22-03-12-Alpine.js and Tailwind HTML setup.md"
+	},
+	{
+		"title": "Web Scraping Javascript with Python",
+		"slug": "web-scraping-javascript-with-python-xhr",
+		"slug_url": "/blog/web-scraping-javascript-with-python-xhr",
+		"path": "blog/2020/16032020-python-xhr-scraping.md"
+	},
+	{
+		"title": "Design First API Development with Goa",
+		"slug": "Design-First-API-Development-with-Goa",
+		"slug_url": "/blog/design-first-api-development-with-goa",
+		"path": "blog/2024/2024-01-09-Design-First-API-Development-with-Goa.md"
+	}
+]
