+++ layouts/partials/umami.html
@@ -1,0 +1,5 @@
+{{ if ne hugo.Environment "development" }}
+<script defer src="https://{{ .Site.Params.umami.scriptUrl }}/script.js"
+        data-website-id="{{ .Site.Params.umami.websiteId }}">
+</script>
+{{ end }}
