+++ layouts/shortcodes/plausible.html
@@ -9,7 +9,6 @@
         </div>
     </div>
     <script src="https://cdnjs.cloudflare.com/ajax/libs/Chart.js/3.7.0/chart.min.js"></script>
-    <script defer src="https://umami.infra.ptco.rocks/script.js" data-website-id="fec62c79-ce55-458a-8e9a-3e9bcf38c790"></script>
     <script type="application/javascript">
         // Calling the API and rendering the Chart are done inside a single
         // IIFE so that no function name duplicate declaration errors occur.
