+++ scripts/zet/fetch-zets.go
@@ -2,6 +2,7 @@ package main
 
 import (
 	"bufio"
+	"bytes"
 	"encoding/base64"
 	"encoding/json"
 	"errors"
@@ -10,7 +11,10 @@ import (
 	"log"
 	"net/http"
 	"os"
+	"path/filepath"
+	"strconv"
 	"strings"
+	"text/template"
 	"time"
 )
 
@@ -21,6 +25,70 @@ var (
 	contentUrl = fmt.Sprintf("%s/repos/%s/%s/contents", gh, owner, repo)
 	readmeUrl  = fmt.Sprintf("%s/repos/%s/%s/readme", gh, owner, repo)
 )
+
+type Content struct {
+	Name        string      `json:"name"`
+	Path        string      `json:"path"`
+	Sha         string      `json:"sha"`
+	Size        int         `json:"size"`
+	Url         string      `json:"url"`
+	HtmlUrl     string      `json:"html_url"`
+	GitUrl      string      `json:"git_url"`
+	DownloadUrl interface{} `json:"download_url"`
+	Type        string      `json:"type"`
+	Links       struct {
+		Self string `json:"self"`
+		Git  string `json:"git"`
+		Html string `json:"html"`
+	} `json:"_links"`
+}
+type Readme struct {
+	Name        string    `json:"name"`
+	Path        string    `json:"path"`
+	Sha         string    `json:"sha"`
+	Size        int       `json:"size"`
+	Url         string    `json:"url"`
+	HtmlUrl     string    `json:"html_url"`
+	GitUrl      string    `json:"git_url"`
+	DownloadUrl string    `json:"download_url"`
+	Type        string    `json:"type"`
+	Content     string    `json:"content"`
+	Encoding    string    `json:"encoding"`
+	Date        time.Time `json:"date"`
+	Links       struct {
+		Self string `json:"self"`
+		Git  string `json:"git"`
+		Html string `json:"html"`
+	} `json:"_links"`
+}
+
+type ZetJson struct {
+	Name        string    `json:"name"`
+	Path        string    `json:"path"`
+	Sha         string    `json:"sha"`
+	Size        int       `json:"size"`
+	Url         string    `json:"url"`
+	HtmlUrl     string    `json:"html_url"`
+	GitUrl      string    `json:"git_url"`
+	DownloadUrl string    `json:"download_url"`
+	Type        string    `json:"type"`
+	Content     string    `json:"content"`
+	Encoding    string    `json:"encoding"`
+	Date        time.Time `json:"date"`
+	Links       struct {
+		Self string `json:"self"`
+		Git  string `json:"git"`
+		Html string `json:"html"`
+	} `json:"_links"`
+}
+
+type ZetTemplate struct {
+	Title  string
+	Slug   string
+	Date   time.Time
+	Body   string
+	IsoSec string
+}
 
 func main() {
 	log.Println("retrieving zets from github")
@@ -179,7 +247,7 @@ func parseDate(r *Readme) time.Time {
 // if the current file is not equal to the number of zets found in the zet repo on GitHub.
 // This prevents needless commits via GitHub actions.
 func noNewZets(r []*Readme) bool {
-	z, err := os.Stat("zet.json")
+	z, err := os.Stat("./assets/zet.json")
 	if err != nil {
 		log.Println("zet.json does not exist", err)
 		return false
@@ -213,12 +281,11 @@ func noNewZets(r []*Readme) bool {
 }
 
 func writeJSONToFile(s []*Readme) error {
-
 	j, err := json.MarshalIndent(s, "", "\t")
 	if err != nil {
 		return err
 	}
-	err = os.WriteFile("zet.json", j, 0644)
+	err = os.WriteFile("./assets/zet.json", j, 0644)
 	if err != nil {
 		return err
 	}
@@ -252,7 +319,10 @@ func decodeJSONBody(r *http.Response, dst interface{}) error {
 		// *json.SyntaxError. If it does, then return a user-readable error
 		// message including the location of the problem
 		case errors.As(err, &syntaxError):
-			return fmt.Errorf("body contains badly-formed JSON (at character %d)", syntaxError.Offset)
+			return fmt.Errorf(
+				"body contains badly-formed JSON (at character %d)",
+				syntaxError.Offset,
+			)
 
 		// Decode() can also return an io.ErrUnexpectedEOF for JSON syntax errors. This is
 		// checked for with errors.Is() and returns a generic error message to the client.
@@ -263,9 +333,15 @@ func decodeJSONBody(r *http.Response, dst interface{}) error {
 		// struct.
 		case errors.As(err, &unmarshallTypeError):
 			if unmarshallTypeError.Field != "" {
-				return fmt.Errorf("body contains incorrect JSON type for field %q", unmarshallTypeError.Field)
+				return fmt.Errorf(
+					"body contains incorrect JSON type for field %q",
+					unmarshallTypeError.Field,
+				)
 			}
-			return fmt.Errorf("body contains incorrect JSON type (at character %d)", unmarshallTypeError.Offset)
+			return fmt.Errorf(
+				"body contains incorrect JSON type (at character %d)",
+				unmarshallTypeError.Offset,
+			)
 
 		// An EOF error will be returned by Decode() if the request body is empty. Use errors.Is()
 		// to check for this and return a human-readable error message
@@ -307,38 +383,153 @@ func decodeJSONBody(r *http.Response, dst interface{}) error {
 	return nil
 }
 
-type Content struct {
-	Name        string      `json:"name"`
-	Path        string      `json:"path"`
-	Sha         string      `json:"sha"`
-	Size        int         `json:"size"`
-	Url         string      `json:"url"`
-	HtmlUrl     string      `json:"html_url"`
-	GitUrl      string      `json:"git_url"`
-	DownloadUrl interface{} `json:"download_url"`
-	Type        string      `json:"type"`
-	Links       struct {
-		Self string `json:"self"`
-		Git  string `json:"git"`
-		Html string `json:"html"`
-	} `json:"_links"`
+func createZetMarkdownFiles() {
+	p, _ := os.Getwd()
+	zetPath := filepath.Join(p, "content/zet")
+
+	zetFile, err := os.ReadFile("./assets/zet.json")
+	if err != nil {
+		log.Fatalln("failed to parse zet.json")
+	}
+
+	var zets []ZetJson
+	err = json.Unmarshal(zetFile, &zets)
+
+	// Get years and create directory paths for them if not exist.
+	yrs := getZetYears(zets)
+	for _, v := range yrs {
+		v := strconv.Itoa(v)
+		exist, err := exists(filepath.Join(zetPath, v))
+		if err != nil {
+			log.Fatalln(err)
+		}
+		if !exist {
+			_ = os.MkdirAll(filepath.Join(zetPath, v), os.ModePerm)
+			log.Printf("created %q directory", filepath.Join(zetPath, v))
+		}
+	}
+
+	for _, z := range zets[:] {
+		zbody, err := getZetContents(z)
+		if err != nil {
+			log.Fatalln("failed to retrieve zet body from github", err)
+		}
+		zt := ZetTemplate{
+			Title:  z.Content,
+			Slug:   slugify(z.Content),
+			Date:   z.Date,
+			Body:   string(zbody),
+			IsoSec: z.Path,
+		}
+
+		_, err = createTemplate(&zt)
+		if err != nil {
+			log.Fatalln("failed to create template", err)
+		}
+
+		zetYear := strconv.Itoa(z.Date.Year())
+		writeZetMarkdown(zt, filepath.Join(zetPath, zetYear))
+
+	}
 }
-type Readme struct {
-	Name        string    `json:"name"`
-	Path        string    `json:"path"`
-	Sha         string    `json:"sha"`
-	Size        int       `json:"size"`
-	Url         string    `json:"url"`
-	HtmlUrl     string    `json:"html_url"`
-	GitUrl      string    `json:"git_url"`
-	DownloadUrl string    `json:"download_url"`
-	Type        string    `json:"type"`
-	Content     string    `json:"content"`
-	Encoding    string    `json:"encoding"`
-	Date        time.Time `json:"date"`
-	Links       struct {
-		Self string `json:"self"`
-		Git  string `json:"git"`
-		Html string `json:"html"`
-	} `json:"_links"`
+
+func exists(path string) (bool, error) {
+	_, err := os.Stat(path)
+	if err == nil {
+		return true, nil
+	}
+	if os.IsNotExist(err) {
+		return false, nil
+	}
+	return false, err
+}
+
+// getZetYears extracts only unique years from an array of all zets
+func getZetYears(zets []ZetJson) []int {
+	var yrs []int
+	for _, y := range zets {
+		yrs = append(yrs, y.Date.Year())
+	}
+
+	return removeDupeYears(yrs)
+}
+
+// removeDupeYears whittles down an array of years as int's to a list of unqiue
+// years.
+func removeDupeYears(arr []int) []int {
+	keys := make(map[int]bool)
+	var l []int
+	for _, item := range arr {
+		if _, value := keys[item]; !value {
+			keys[item] = true
+			l = append(l, item)
+		}
+	}
+	return l
+}
+
+func getZetContents(z ZetJson) ([]byte, error) {
+	cl := http.Client{Timeout: 10 * time.Second}
+	req, err := http.NewRequest(http.MethodGet, z.DownloadUrl, nil)
+	if err != nil {
+		return nil, err
+	}
+	res, err := cl.Do(req)
+	if err != nil {
+		return nil, err
+	}
+
+	if res.Body != nil {
+		defer res.Body.Close()
+	}
+
+	body, err := io.ReadAll(res.Body)
+	if err != nil {
+		return nil, err
+	}
+	return body, nil
+}
+
+func writeZetMarkdown(zt ZetTemplate, path string) {
+	tpl, err := createTemplate(&zt)
+	if err != nil {
+		log.Fatalln(err)
+	}
+
+	err = os.WriteFile(fmt.Sprintf("%s/%s-%s.md", path, zt.IsoSec, zt.Slug), tpl, 0644)
+	if err != nil {
+		log.Fatalln(err)
+	}
+	log.Printf("wrote %q markdown file into zet directory", zt.Title)
+}
+
+func slugify(s string) string {
+	return strings.Replace(s, " ", "-", -1)
+}
+
+func createTemplate(z *ZetTemplate) ([]byte, error) {
+	t := template.New("zet")
+
+	parse, err := t.Parse(strings.TrimSpace(`
++++
+title = "{{.Title}}"
+categories = ["zet"]
+tags = ["zet"]
+slug = "{{.Slug}}"
+date = "{{.Date}}"
+draft = "false"
+ShowToc = "true"
++++
+
+{{.Body}}
+	`))
+	if err != nil {
+		return nil, err
+	}
+	var buf bytes.Buffer
+	err = parse.Execute(&buf, z)
+	if err != nil {
+		return nil, err
+	}
+	return buf.Bytes(), nil
 }
